Redirect to request.fullpath instead of request.path
Imagine your application is an OAuth provider and application ABC
will use you as the authentication service.  From ABC, the user will
get redirected to `oauth/authorize` with the needed params (`client_id`,
`scope`, and others). If the user is not logged in, the provider will
show the login form and after a successful login, the user should be
redirected to `/ouath/authorize?client_id=123&scope=public` and not
to `/oauth/authorize`, since this route is invalid.